[DEV] Add Upload to remote server
This commits adds the ground work to allow the upload of an Argumentation system to the remote server.
A new Toolbar Component as been added to the online workspace to expose such operation. Said toolbar expose events that the workspace component can wire onto to execute the code required to upload the argumentation system.
Bootstrap has been updated to version 4 and the ng-bootstrap package has been added to allow the display of modal to ask user confirmation before potentially unsafe operation.
Known error : ng-boostrap emit a warning during npm install because it require an Angular 4 peer dependency. Until the package is updated the warning will stay.
Bootstrap tooltip require tether.