[DEV] Refactor Modal Form for pattern creation to be standalone component
This commits aims to change the way forms are created and assigned to modals.
Now modals should be stand alone component that can be opened by passing a
reference to the Component name. This allows for a greater flexibility in testing
and less cluttering in the workspace HTML template code. Other modal should be
refactored to be the same.